% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rdir_maxent}
\alias{rdir_maxent}
\title{Generate random numbers from a Dirichlet distribution via Maximum Entropy}
\usage{
rdir_maxent(n, shares, ...)
}
\arguments{
\item{n}{Integer. Number of random vectors to generate.}

\item{shares}{Numeric vector. A best-guess (mean) of the shares. Must sum to 1.}

\item{...}{Other parameters passed on to \code{\link[=find_gamma_maxent]{find_gamma_maxent()}}.}
}
\value{
A numeric matrix of dimension \code{n} times \code{length(shares)}, where
each row is a random deviate from the Dirichlet distribution.
}
\description{
This function estimates the concentration parameter \eqn{\gamma} (sometimes noted
\eqn{\alpha_0}) by applying the maximum entropy principle via
\code{\link[=find_gamma_maxent]{find_gamma_maxent()}}. The resulting parameter is then used to generate random
vectors from a Dirichlet distribution.
}
\details{
For details on how \eqn{\gamma} is fitted, see \code{\link[=find_gamma_maxent2]{find_gamma_maxent2()}} or the
references therein. The entropy is evaluated using \code{\link[=dirichlet_entropy]{dirichlet_entropy()}}.
}
\examples{
rdir_maxent(5, c(0.2, 0.3, 0.5))
}
\seealso{
\code{\link[=find_gamma_maxent]{find_gamma_maxent()}}, \code{\link[=dirichlet_entropy]{dirichlet_entropy()}}, \code{\link[=rdirichlet_standard]{rdirichlet_standard()}}
}
