% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rdirichlet_uniform}
\alias{rdirichlet_uniform}
\title{Generate uniform random numbers from a Dirichlet distribution}
\usage{
rdirichlet_uniform(n, length, names = NULL)
}
\arguments{
\item{n}{Integer. Number of random vectors to generate.}

\item{length}{Integer. The number of variables (columns) in the output.}

\item{names}{Optional character vector of length \code{length} giving column names.}
}
\value{
A numeric matrix of dimension \code{n} times \code{length}, with each row
a Dirichlet random deviate drawn from a uniform (each \eqn{\alpha=1}) distribution.
}
\description{
Generates random vectors from a Dirichlet distribution with each \eqn{\alpha = 1}.
This implies a uniform distribution over the simplex.
}
\details{
For details, see
\href{https://en.wikipedia.org/wiki/Dirichlet_distribution#When_each_alpha_is_1}{Wikipedia on the Dirichlet distribution}.
}
\examples{
rdirichlet_uniform(5, 3)
}
\seealso{
\code{\link[=rdirichlet_standard]{rdirichlet_standard()}}
}
