% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_maxent.R
\name{dirichlet_entropy_grad}
\alias{dirichlet_entropy_grad}
\title{Derivative of the Dirichlet entropy function.
Derivated using the \code{Deriv} package and \code{autodiffr}.
For derivation see script:}
\usage{
dirichlet_entropy_grad(x, shares)
}
\arguments{
\item{shares}{}
}
\description{
NOTE: Derivation does not seem to be correct. Is not used anymore in the optimisation.
}
