% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshares.R
\name{rdirichlet_standard}
\alias{rdirichlet_standard}
\title{Functions to generate random deviates from the Dirichlet distribution.}
\usage{
rdirichlet_standard(n, alpha, threshold = 0.01)
}
\arguments{
\item{threshold}{}
}
\description{
Like `gtools::rdirichlet`, but with an adjustment when `alpha` contains small
values below a given `threshold`.
}
\details{
Basically the same as `rdirichlet`, but keeping the parametrisation of `gtools::rdirichlet`
to allow easy switching between the two.

Details see `?rdirichlet`.
}
