% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragg.R
\name{rlnorm2}
\alias{rlnorm2}
\title{Random number generation for the lognormal distribution parameterised to the `mean` and `sd` on the **linear** scale.
Taken from: https://stackoverflow.com/questions/56821688/sample-a-lognormal-distribution-to-an-exact-mean-and-sd}
\usage{
rlnorm2(n, mean, sd)
}
\arguments{
\item{mean}{the mean on the linear scale}

\item{sd}{the standard deviation on the linear scale}
}
\description{
Random number generation for the lognormal distribution parameterised to the `mean` and `sd` on the **linear** scale.
Taken from: https://stackoverflow.com/questions/56821688/sample-a-lognormal-distribution-to-an-exact-mean-and-sd
}
